requirements.txt
grpcio-tools==1.35.0
kubernetes
jinja2
pyyaml
